feat(helius_api): add max_rate_limit_retries configuration
- Move max_rate_limit_retries from a constant to a configurable parameter
- Update HeliusApi struct and new function to include max_rate_limit_retries
- Replace hardcoded MAX_RATE_LIMIT_RETRIES with configurable value
- Update related tests to use the new configuration parameter

+++ src/settings.rs
@@ -82,7 +82,7 @@ pub struct RpcSettings {
     pub url: String,
     pub rps: u32,
     pub max_concurrent: usize,
-    pub max_rate_limit_retries: u32,
+    pub max_rate_limit_retries: usize,
 }
 
 #[derive(Debug, Deserialize, Clone)]
